update my fork

+++ FWCore/Skeletons/python/pkg.py
@@ -8,6 +8,7 @@ Author     : Valentin Kuznetsov <vkuznet@gmail.com>
 Description: AbstractGenerator class provides basic functionality
 to generate CMSSW class from given template
 """
+from __future__ import print_function
 
 # system modules
 import os
@@ -48,6 +49,7 @@ class AbstractPkg(object):
         self.author = user_info(self.config.get('author', None))
         self.date   = time.strftime("%a, %d %b %Y %H:%M:%S GMT", time.gmtime())
         self.not_in_dir = self.config.get('not_in_dir', [])
+        self.working_dir = self.config.get('working_dir')
 
     def tmpl_etags(self):
         "Scan template files and return example tags"
@@ -69,7 +71,7 @@ class AbstractPkg(object):
     def print_etags(self):
         "Print out template example tags"
         for key in self.tmpl_etags():
-            print key
+            print(key)
 
     def tmpl_tags(self):
         "Scan template files and return template tags"
@@ -90,7 +92,7 @@ class AbstractPkg(object):
     def print_tags(self):
         "Print out template keys"
         for key in self.tmpl_tags():
-            print key
+            print(key)
 
     def parse_etags(self, line):
         """
@@ -113,6 +115,8 @@ class AbstractPkg(object):
         "Create new file from given template name and set of arguments"
         code = ""
         read_code = False
+        if os.path.exists(fname):
+            return
         with open(fname, 'w') as stream:
             for line in open(tmpl_name, 'r').readlines():
                 line = self.parse_etags(line)
@@ -132,7 +136,7 @@ class AbstractPkg(object):
                     continue
                 if  not read_code:
                     for key, val in kwds.items():
-                        if  isinstance(val, basestring):
+                        if  isinstance(val, str):
                             line = line.replace(key, val)
                     stream.write(line)
 
@@ -149,7 +153,7 @@ class AbstractPkg(object):
         args = self.config.get('args', None)
         kwds.update(args)
         if  self.debug:
-            print "Template tags:"
+            print("Template tags:")
             pprint.pprint(kwds)
         return kwds
 
@@ -170,7 +174,7 @@ class AbstractPkg(object):
             if  os.path.isdir(self.pname):
                 msg  = "Can't create package '%s'\n" % self.pname
                 msg += "Directory %s is already exists" % self.pname
-                print msg
+                print(msg)
                 sys.exit(1)
             os.makedirs(self.pname)
             os.chdir(self.pname)
@@ -196,7 +200,7 @@ class AbstractPkg(object):
                 kwds = self.get_kwds()
                 if  not sources:
                     msg = 'Unable to find skeleton for extension "%s"' % ext
-                    print msg
+                    print(msg)
                     sys.exit(1)
             bdir = os.environ.get('CMSSW_BASE', '')
             dirs = os.getcwd().replace(bdir, '').split('/')
@@ -225,9 +229,12 @@ class AbstractPkg(object):
                 fname, ext = os.path.splitext(src)
                 if  tmpl_files != ext:
                     continue
+                #also reject if this is the wrong directory
+                if self.working_dir and src.split('/')[-2] != self.working_dir:
+                    continue
                 src = src.split('/')[-1]
             if  self.debug:
-                print "Read", src
+                print("Read", src)
             items = src.split('/')
             if  items[-1] == '/':
                 items = items[:-1]
@@ -238,11 +245,13 @@ class AbstractPkg(object):
             else:
                 ftype = 'dir'
             name2gen  = src # new file we'll create
-            if  tname.split('.')[0] == self.tmpl: # need to substitute
+            if items[-1] == 'testBuildFile.xml':
+              name2gen = '/'.join(src.split('/')[:-1])+'/BuildFile.xml'
+            if  -1 !=tname.split('.')[0].find(self.tmpl): # need to substitute
                 name2gen  = name2gen.replace(self.tmpl, self.pname)
             name2gen  = os.path.join(os.getcwd(), name2gen)
             if  self.debug:
-                print "Create", name2gen
+                print("Create", name2gen)
             if  ftype == 'dir':
                 if  not os.path.isdir(name2gen):
                     os.makedirs(name2gen)
@@ -259,7 +268,7 @@ class AbstractPkg(object):
             msg = 'Generated %s file' % ', '.join(gen_files)
             if  len(gen_files) > 1:
                 msg += 's'
-        print msg
+        print(msg)
         # return back where we started
         os.chdir(cdir)
         if  msg.find('New package') != -1:
